- Finished inventory item loading

+++ trunk/Rouge/RougeLike/Character/ArmorDefinition.cs
@@ -5,8 +5,8 @@ namespace RougeLike.Character
 {
   public class ArmorDefinition : InventoryItemDefinition
   {
-    public int BeamRating { get; set; }
-    public int ProjectileRating { get; set; }
+    public float BeamRating { get; set; }
+    public float ProjectileRating { get; set; }
     public SkillSet Buffs { get; set; }
     public List<WeaponEffect> Immunities { get; set; } // Must be checked by weapon effect before effect is applied
   }
