+++ trunk/Rouge/RougeLike/Character/WeaponDefinition.cs
@@ -6,10 +6,10 @@ namespace RougeLike.Character
   public class WeaponDefinition : InventoryItemDefinition
   {
     public WeaponType WeaponType { get; set; }
-    public int RateOfFire { get; set; }
-    public int Damage { get; set; }
-    public int ProjectileSpeed { get; set; }
-    public int ProjectileSize { get; set; }
+    public float RateOfFire { get; set; }
+    public float Damage { get; set; }
+    public float ProjectileSpeed { get; set; }
+    public float ProjectileSize { get; set; }
     public bool Continuous { get; set; }
     public WeaponEffect Effect { get; set; } //Is applied when shot hits another character
     public string MuzzleFlashAnim { get; set; }
